Leave early if size parameter is 0 or less
This saves us a lot of work if it is possible for the size to be empty, and also so we do not write to any memory if somehow we get a 0 size buffer.

+++ ppdc/ppdc.h
@@ -509,7 +509,7 @@ class ppdcSource			//// Source File
   ppdcChoice	*get_resolution(ppdcFile *fp);
   ppdcProfile	*get_simple_profile(ppdcFile *fp);
   ppdcMediaSize	*get_size(ppdcFile *fp);
-  char		*get_token(ppdcFile *fp, char *buffer, int buflen);
+  char		*get_token(ppdcFile *fp, char *buffer, size_t buflen);
   ppdcVariable	*get_variable(ppdcFile *fp);
   int		import_ppd(const char *f);
   int		quotef(cups_file_t *fp, const char *format, ...);
